Made some changes to a bunch of classes to get it to compile
Bugfix: if containsHeader was set to false the statement to fill the header row wouldn't be executed.
Refactor: changed raw arrays to collection types.
Refactor: FeatureClass now keeps track of whether or not it's an expected output.
Refactor: Added a NumericFeatureClass subclass specifically for dealing with numeric types.
Refactor: Added a Preprocessor superclass for preprocessors.
Bugfix: Objectives and chromosomes now use the new input system. It probably works, but please add a regression suite.
Misc: commented out the code in the GUI workspace controller that enabled data normalisation. We're moving to preprocessors so we're better off rewriting it instead.
Warning: the changes to Objectives and Chromosomes are temporary workarounds to enable testing, they'll need to be changed later on.