Add a combo box to allow configurations to select different datasets
Refactor: moves the primitives inside of SearchExecutor into SearchConfigurationModel. Additional work will need to be done to handle toggling them.
Feature: controls in the configuration views now have a bit of space between them.
Feature: the dataset used by a particular configuration can now be changed visually. This still needs to be tied in to the back end.