Refactor the CLI to use graphing and file export utilities
Feature: the user can now specify if they want to export a graph of the results and/or export it to a CSV file.
Feature: the CLI will now print a README in a structured directory that lists most of the arguments passed to it.
Comment: the CLI is hard coded to use SEAMO until the InstrumenterFactory class is complete.